move main files
