+++ src/connect.js
@@ -1,13 +1,14 @@
-import { render } from './dom.js';
+import { render } from '../node_modules/lit-html/lit-html.js';
 import Base from './base.js';
 
-export default function (name, component, lifecycle = {}) {
+export default function (name, component, store, lifecycle = {}) {
   customElements.define(
     name,
     class extends Base {
       constructor () {
         super();
         this.lifecycle = lifecycle;
+        store.subscribe(this.doRender.bind(this));
         this.doLifecycleFunc('created');
       }
 
@@ -17,7 +18,7 @@ export default function (name, component, lifecycle = {}) {
       }
 
       render () {
-        const res = component(this.props);
+        const res = component(store)(this.props);
         render(res, this.shadowRoot);
       }
     }
