+++ src/stateful.js
@@ -1,4 +1,4 @@
-import { render } from './dom.js';
+import { render } from '../node_modules/lit-html/lit-html.js';
 import Base from './base.js';
 import { getNextState, getState } from './helpers.js';
